feat(util): rename @svrx/svrx-util to @svrx/util
