+++ packages/svrx-util/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@svrx/svrx-util",
+  "name": "@svrx/util",
   "version": "1.0.0",
   "description": "util package of svrx",
   "homepage": "https://svrx.io/",
